+++ src/extract.rs
@@ -28,9 +28,8 @@ let rules: &[Rewrite<SimpleLanguage, ()>] = &[
 ];
 
 let start = "(+ 0 (* 1 10))".parse().unwrap();
-let (mut egraph, root) = EGraph::from_expr(&start);
-
-Runner::default().run(&mut egraph, &rules);
+let runner = Runner::new().with_expr(&start).run(&rules);
+let (egraph, root) = (runner.egraph, runner.roots[0]);
 
 let mut extractor = Extractor::new(&egraph, AstSize);
 let (best_cost, best) = extractor.find_best(root);