+++ src/lib.rs
@@ -1,4 +1,4 @@
-// #![warn(missing_docs)]
+#![warn(missing_docs)]
 /*!
 [`EGraph`]s (and almost everything else in this crate) are
 parameterized over the language given by the user (by implementing
@@ -46,11 +46,11 @@ let rules: &[Rewrite<SimpleLanguage, ()>] = &[
 ];
 
 let start = "(+ 0 (* 1 foo))".parse().unwrap();
-let (egraph, report) = Runner::default().run_expr(start, &rules);
+let runner = Runner::new().with_expr(&start).run(&rules);
 println!(
     "Stopped after {} iterations, reason: {:?}",
-    report.iterations.len(),
-    report.stop_reason
+    runner.iterations.len(),
+    runner.stop_reason
 );
 ```
 !*/